one area entity and dto
