+++ backend/src/main/java/ru/hh/career/solution/config/BeanConfig.java
@@ -3,8 +3,8 @@ package ru.hh.career.solution.config;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
-import ru.hh.career.solution.area.dto.AreaParser;
-import ru.hh.career.solution.area.service.ParserAreaService;
+import ru.hh.career.solution.parser.AreaParser;
+import ru.hh.career.solution.service.ParserAreaService;
 import ru.hh.career.solution.dao.AccountDao;
 import ru.hh.career.solution.dao.AdviserDao;
 import ru.hh.career.solution.dao.CheckerOnNullDao;
