+++ backend/src/main/java/ru/hh/career/solution/mapper/AreaMapper.java
@@ -7,13 +7,15 @@ public class AreaMapper {
 
   public static Area mapToArea(AreaDto areaDto) {
     return areaDto == null ? null : new Area(
-      areaDto.getAreaId(),
-      areaDto.getName());
+      areaDto.getId(),
+      areaDto.getName(),
+      areaDto.getParentId());
   }
 
   public static AreaDto mapToAreaDto(Area area) {
     return area == null ? null : new AreaDto(
-      area.getAreaId(),
-      area.getName());
+      area.getId(),
+      area.getName(),
+      area.getParentId());
   }
 }
